app_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
PyMySQL
alembic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings # Para gestionar la configuración desde .env
pyotp # <--- AÑADIDO PARA 2FA